Moved Link logic to a Mixin. This will make it easier to create custom Link components.
