+++ lib/LinkMixin.js
@@ -1,0 +1,33 @@
+var React               = require('react');
+var invariant           = require('react/lib/invariant');
+var RoutingContextMixin = require('./RoutingContextMixin');
+var makeHref            = require('./makeHref');
+
+var LinkMixin = {
+  mixins: [RoutingContextMixin],
+
+  propTypes: {
+    to: React.PropTypes.string,
+    href: React.PropTypes.string
+  },
+
+  onClick: function(e) {
+    e.preventDefault();
+    this.navigate(this.href());
+  },
+
+  href: function() {
+    if (this.props.href) {
+      return this.props.href;
+    } else if (this.props.to) {
+      return this.getRouting().makeHref(this.props.to, this.props);
+    } else {
+      invariant(
+        false,
+        'provide either "to" or "href" prop to Link component'
+      );
+    }
+  }
+}
+
+module.exports = LinkMixin;
